kaifawanc

+++ app/src/main/java/com/dafa/qipai/dafaqipai/util/ApiUtils.java
@@ -241,7 +241,7 @@ public class ApiUtils {
             Integer pageIndex,
             Integer pageSize,
             Integer status,
-            Long type,
+            Integer type,
             OkGoCallBack okGoCallBack
     ) {
         Map<String, Object> paramsMap = new HashMap<>();
